client_sdk/requirements.txt
pynacl==1.5.0
requests==2.31.0
blake3==0.4.1
